requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
streamlit==1.30.0
requests==2.31.0
python-multipart==0.0.6
firebase-admin==6.4.0
dnspython==2.4.2
